requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.12.0
pydub==0.25.1
python-multipart==0.0.6
pytz==2023.3